RandomVariables reorganization
